+++ jablib/src/main/java/org/jabref/logic/FilePreferences.java
@@ -32,6 +32,7 @@ public class FilePreferences {
     private final StringProperty fileDirectoryPattern = new SimpleStringProperty();
     private final BooleanProperty downloadLinkedFiles = new SimpleBooleanProperty();
     private final BooleanProperty fulltextIndexLinkedFiles = new SimpleBooleanProperty();
+    private final BooleanProperty fulltextIndexLinkedFilesShouldCheckForModifications = new SimpleBooleanProperty();
     private final ObjectProperty<Path> workingDirectory = new SimpleObjectProperty<>();
     private final BooleanProperty createBackup = new SimpleBooleanProperty();
     private final ObjectProperty<Path> backupDirectory = new SimpleObjectProperty<>();
@@ -57,6 +58,7 @@ public class FilePreferences {
                 "",                                  // fileDirectoryPattern
                 true,                                // downloadLinkedFiles
                 true,                                // fulltextIndexLinkedFiles
+                false,                               // fulltextIndexLinkedFilesShouldCheckForModifications
                 Directories.getUserDirectory(),      // workingDirectory
                 true,                                // createBackup
                 Directories.getBackupDirectory(),    // backupDirectory
@@ -80,6 +82,7 @@ public class FilePreferences {
                            String fileDirectoryPattern,
                            boolean downloadLinkedFiles,
                            boolean fulltextIndexLinkedFiles,
+                           boolean fulltextIndexLinkedFilesShouldCheckForModifications,
                            Path workingDirectory,
                            boolean createBackup,
                            Path backupDirectory,
@@ -100,6 +103,7 @@ public class FilePreferences {
         this.fileDirectoryPattern.setValue(fileDirectoryPattern);
         this.downloadLinkedFiles.setValue(downloadLinkedFiles);
         this.fulltextIndexLinkedFiles.setValue(fulltextIndexLinkedFiles);
+        this.fulltextIndexLinkedFilesShouldCheckForModifications.setValue(fulltextIndexLinkedFilesShouldCheckForModifications);
         this.workingDirectory.setValue(workingDirectory);
         this.createBackup.setValue(createBackup);
         this.backupDirectory.setValue(backupDirectory);
@@ -204,6 +208,18 @@ public class FilePreferences {
 
     public void setFulltextIndexLinkedFiles(boolean shouldFulltextIndexLinkedFiles) {
         this.fulltextIndexLinkedFiles.set(shouldFulltextIndexLinkedFiles);
+    }
+
+    public boolean shouldFulltextIndexLinkedFilesShouldCheckForModifications() {
+        return fulltextIndexLinkedFilesShouldCheckForModifications.get();
+    }
+
+    public BooleanProperty fulltextIndexLinkedFilesShouldCheckForModifications() {
+        return fulltextIndexLinkedFilesShouldCheckForModifications;
+    }
+
+    public void setFulltextIndexLinkedFilesShouldCheckForModifications(boolean shouldFulltextIndexLinkedFilesShouldCheckForModifications) {
+        this.fulltextIndexLinkedFilesShouldCheckForModifications.set(shouldFulltextIndexLinkedFilesShouldCheckForModifications);
     }
 
     public Path getWorkingDirectory() {